+++ sdk/keyvault/azure-keyvault-certificates/samples/list_operations.py
@@ -25,7 +25,7 @@ from azure.core.exceptions import HttpResponseError
 #
 # 2. List certificates from the Key Vault (list_properties_of_certificates)
 #
-# 3. List certificate versions from the Key Vault (list_certificate_versions)
+# 3. List certificate versions from the Key Vault (list_properties_of_certificate_versions)
 #
 # 4. List deleted certificates from the Key Vault (list_deleted_certificates). The vault has to be soft-delete enabled
 # to perform this operation.
@@ -82,7 +82,7 @@ try:
     # You need to check all the different tags your bank account certificate had previously. Let's print
     # all the versions of this certificate.
     print("\n.. List versions of the certificate using its name")
-    certificate_versions = client.list_certificate_versions(bank_cert_name)
+    certificate_versions = client.list_properties_of_certificate_versions(bank_cert_name)
     for certificate_version in certificate_versions:
         print(
             "Bank Certificate with name '{0}' with version '{1}' has tags: '{2}'.".format(