+++ sdk/keyvault/azure-keyvault-certificates/tests/test_examples_certificates.py
@@ -145,17 +145,17 @@ class TestExamplesKeyVault(KeyVaultTestCase):
             print(certificate.enabled)
 
         # [END list_certificates]
-        # [START list_certificate_versions]
+        # [START list_properties_of_certificate_versions]
 
         # get an iterator of a certificate's versions
-        certificate_versions = certificate_client.list_certificate_versions(name="certificate-name")
+        certificate_versions = certificate_client.list_properties_of_certificate_versions(name="certificate-name")
 
         for certificate in certificate_versions:
             print(certificate.id)
             print(certificate.updated_on)
             print(certificate.version)
 
-        # [END list_certificate_versions]
+        # [END list_properties_of_certificate_versions]
         # [START list_deleted_certificates]
 
         # get an iterator of deleted certificates (requires soft-delete enabled for the vault)