+++ sdk/keyvault/azure-keyvault-certificates/HISTORY.md
@@ -9,6 +9,7 @@
 - Renamed `get_certificate_with_policy` to `get_certificate`
 - Renamed `list_certificates` to `list_properties_of_certificates`
 - Renamed `list_issuers` to `list_properties_of_issuers`
+- Renamed `list_certificate_versions` to `list_properties_of_certificate_versions`
 - `create_certificate` now has policy as a required parameter
 - All optional positional parameters besides `version` have been moved to kwargs